Split detail window code into a separate file
